+++ app/src/main/java/com/napzak/market/data/marketinfo/datasource/MarketInfoDataSource.kt
@@ -2,10 +2,6 @@ package com.napzak.market.data.marketinfo.datasource
 
 import com.napzak.market.core.network.BaseResponse
 import com.napzak.market.data.marketinfo.dto.MarketInfoResponse
-import com.napzak.market.data.marketinfo.dto.MarketProductBuyItemsRequest
-import com.napzak.market.data.marketinfo.dto.MarketProductBuyItemsResponse
-import com.napzak.market.data.marketinfo.dto.MarketProductSellItemsRequest
-import com.napzak.market.data.marketinfo.dto.MarketProductSellItemsResponse
 import com.napzak.market.data.marketinfo.service.MarketInfoService
 import javax.inject.Inject
 
@@ -14,21 +10,4 @@ class MarketInfoDataSource @Inject constructor(
 ) {
     suspend fun getMarketInfo(request: Long): BaseResponse<MarketInfoResponse> =
         service.getMarketInfo(storeId = request)
-
-    suspend fun getMarketProductSellItems(request: MarketProductSellItemsRequest): BaseResponse<MarketProductSellItemsResponse> =
-        service.getMarketProductSellItems(
-            storeId = request.storeId,
-            sortOption = request.sortOption,
-            isOnSale = request.isOnSale,
-            isUnopened = request.isUnopened,
-            genreIds = request.genreIds,
-        )
-
-    suspend fun getMarketProductBuyItems(request: MarketProductBuyItemsRequest): BaseResponse<MarketProductBuyItemsResponse> =
-        service.getMarketProductBuyItems(
-            storeId = request.storeId,
-            sortOption = request.sortOption,
-            isOnSale = request.isOnSale,
-            genreIds = request.genreIds,
-        )
 }
